refactor: Move PreferencesState into SettingsScreen viewModel package
#222
