+++ app/src/main/java/org/nsh07/wikireader/ui/homeScreen/AppSearchBar.kt
@@ -104,11 +104,11 @@ import org.nsh07.wikireader.data.WikiPrefixSearchResult
 import org.nsh07.wikireader.data.WikiSearchResult
 import org.nsh07.wikireader.ui.image.FeedImage
 import org.nsh07.wikireader.ui.settingsScreen.LanguageBottomSheet
+import org.nsh07.wikireader.ui.settingsScreen.viewModel.PreferencesState
 import org.nsh07.wikireader.ui.theme.WRShapeDefaults.bottomListItemShape
 import org.nsh07.wikireader.ui.theme.WRShapeDefaults.middleListItemShape
 import org.nsh07.wikireader.ui.theme.WRShapeDefaults.topListItemShape
 import org.nsh07.wikireader.ui.viewModel.AppSearchBarState
-import org.nsh07.wikireader.ui.viewModel.PreferencesState
 
 @OptIn(
     ExperimentalMaterial3Api::class,
